# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Merchant Voice API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginEmployee
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeLoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeSession"
        "401":
          description: Invalid employee credentials
        "429":
          description: Too many sign-in attempts
...
  /auth/change-password:
    post:
      operationId: changeEmployeePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Password changed and other sessions revoked
        "400":
          description: Invalid password
        "401":
          description: Current password is incorrect
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      responses:
        "200":
          description: Employees in the administrator's workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedEmployee"
        "403":
          description: Administrator access required
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedEmployee"
        "409":
          description: Employee ID already exists
  /employees/{id}:
    patch:
      operationId: updateEmployeeAccess
      tags: [employees]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeAccessUpdate"
      responses:
        "200":
          description: Employee access updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedEmployee"
        "404":
          description: Employee not found
  /employees/{id}/reset-password:
    post:
      operationId: resetEmployeePassword
      tags: [employees]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetInput"
      responses:
        "204":
          description: Password reset and active sessions revoked
        "404":
          description: Employee not found
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
...
    Employee:
      type: object
      required: [id, employeeId, name, role]
      properties:
        id: { type: integer }
        employeeId: { type: string }
        name: { type: string }
        role: { type: string }
    ManagedEmployee:
      allOf:
        - $ref: "#/components/schemas/Employee"
        - type: object
          required: [active, createdAt]
          properties:
            active: { type: boolean }
            createdAt: { type: string, format: date-time }
    EmployeeSession:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        employee:
          oneOf:
            - $ref: "#/components/schemas/Employee"
            - type: "null"
    EmployeeLoginInput:
      type: object
      required: [employeeId, password]
      properties:
        employeeId:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
          maxLength: 200
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 8, maxLength: 200 }
        newPassword: { type: string, minLength: 8, maxLength: 200 }
    EmployeeInput:
      type: object
      required: [employeeId, name, role, temporaryPassword]
      properties:
        employeeId: { type: string, minLength: 3, maxLength: 50 }
        name: { type: string, minLength: 2, maxLength: 100 }
        role: { type: string, enum: [Administrator, Employee] }
        temporaryPassword: { type: string, minLength: 8, maxLength: 200 }
    EmployeeAccessUpdate:
      type: object
      minProperties: 1
      properties:
        active: { type: boolean }
        role: { type: string, enum: [Administrator, Employee] }
    PasswordResetInput:
      type: object
      required: [temporaryPassword]
      properties:
        temporaryPassword: { type: string, minLength: 8, maxLength: 200 }
...