# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Merchant Voice API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  role: string;
}