# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Merchant Voice API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputRole } from './employeeInputRole';

export interface EmployeeInput {
  /**
     * @minLength 3
     * @maxLength 50
     */
  employeeId: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  role: EmployeeInputRole;
  /**
     * @minLength 8
     * @maxLength 200
     */
  temporaryPassword: string;
}