# lib/api-zod/src/generated/types/employeeSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Merchant Voice API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';

export interface EmployeeSession {
  authenticated: boolean;
  employee?: Employee | null;
}