# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Merchant Voice API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordChangeInput {
  /**
     * @minLength 8
     * @maxLength 200
     */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  newPassword: string;
}